+++ AssetsFlow.Server/AssetsFlow.Journal.All/HsR.Journal.Repository/Services/JournalRepo/IJournalRepository.cs
@@ -7,7 +7,7 @@ namespace HsR.Journal.DataContext
     public interface IJournalRepository
     {
         Task<(IEnumerable<TradeComposite>?, int totalTradesCount)> GetAllTradeCompositesAsync(Guid userId, int pageNumber = 1, int pageSize = 10);
-        Task<(IEnumerable<TradeComposite>?, int totalTradesCount)> GetFilteredTradeCompositesAsync(Guid userId, IEnumerable<FilterDefinition> filters,
+        Task<(IEnumerable<TradeComposite>?, int totalTradesCount)> GetFilteredTradeCompositesAsync(Guid userId, IEnumerable<FilterDefinition>? filters,
                                                                                             int pageNumber = 1, int pageSize = 10);
 
         Task<TradeComposite> AddTradeCompositeAsync(Guid userId);