+++ AssetsFlow.Server/AssetsFlow.Web.Services/Models/UserDataDTO.cs
@@ -1,4 +1,6 @@
-﻿using System;
+﻿using AssetsFlowWeb.Services.Filtering;
+using HsR.Journal.Services;
+using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.Linq;
@@ -15,5 +17,7 @@ namespace AssetsFlowWeb.Services.Models
         public ICollection<string>? SavedSectors { get; set; }
 
         public ICollection<string>? Symbols { get; set; }
+
+        public ICollection<FilterDefinition>? AvailableFilters { get; set; } = FilterDefinitionFactory.Create();
     }
 }