+++ AssetsFlow.Server/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## [0.9.226-b] - 2025-08-..
-- filtering
+- filtering : added filter model, filters factory
+- filtering: added controller route
+- filtering: added to userdata
+- userdata: added repo wrapper for get symbols, get filters
+
 
 ## [0.9.225-b] - 2025-08-18
 - validate net amount on close trade, return reduce element on close